Additional Chaparral updates for large problems

A few additional fixes for reporting sizes missed in MR !24.

See merge request !25

+++ src/packages/chaparral/vflib/VF_MakeMatrixSymmetric.c
@@ -32,8 +32,8 @@
 
 #include "vf.h"
 
-void VF_MatrixStats(int*, int*, int*);
-void VF_MatrixRowStats(VFsparse_array*, int, int*, int*, int*);
+void VF_MatrixStats(int*, uint64_t*, uint64_t*);
+void VF_MatrixRowStats(VFsparse_array*, int, int*, uint64_t*, uint64_t*);
 
 void VF_MakeMatrixSymmetric(int encl, int method, int output)
 {
